Move LocalGame into pojo package
