+++ app/src/main/java/com/ss20/se2/monopoly/network/OnLocalGamesChangedListener.java
@@ -1,5 +1,7 @@
 package com.ss20.se2.monopoly.network;
 
+import com.ss20.se2.monopoly.pojo.LocalGame;
+
 import java.util.List;
 
 public interface OnLocalGamesChangedListener{
